Updating the number format.

+++ models/intermediate/int_sales_margin.sql
@@ -1,6 +1,6 @@
 SELECT 
     a.*
-    ,CAST(b.purchase_price AS NUMERIC) * a.quantity AS purchase_cost
+    ,CAST(b.purchase_price AS FLOAT64) * a.quantity AS purchase_cost
 FROM
     {{ ref('stg_gz_raw_data__sales') }} AS a
 JOIN
